% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculaty.R
\name{multiplicar}
\alias{multiplicar}
\title{Function multiplicar}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{second number.}
}
\description{
This function is used to multiply two values
}
